refactor: 스타일 조회를 위한 styleController 분리
